chore: final refactor
