+++ packages/talos_common_utils/src/lib.rs
@@ -1,1 +1,2 @@
+pub mod back_pressure;
 pub mod env;
