+++ packages/talos_messenger_core/src/services/mod.rs
@@ -1,3 +1,3 @@
 mod inbound_service;
 
-pub use inbound_service::{MessengerInboundService, MessengerInboundServiceConfig, TalosBackPressureConfig};
+pub use inbound_service::{MessengerInboundService, MessengerInboundServiceConfig};
